# Reduce long[] array allocation for bitset in readBitSetIterator